fixed bugs

+++ script.js
@@ -1,19 +1,13 @@
-let colorArray = ['#FF6633', '#ff8787', '#FF33FF', '#FFFF99', '#00B3E6', 
-		  '#f5c347', '#4dffea', '#ffc45e', '#99FF99', '#f26363',
-		  '#c3f24e', '#809900', '#61bf47', '#db8fac', '#b8e86f', 
-		  '#5bfafc', '#CCFF1A', '#FF1A66', '#ffa599', '#33FFCC',
-		  '#66994D', '#f5c6b3', '#4D8000', '#B33300', '#CC80CC', 
-		  '#eed4ff', '#991AFF', '#E666FF', '#4DB3FF', '#1AB399',
-		  '#E666B3', '#33991A', '#CC9999', '#B3B31A', '#00E680', 
-		  '#f2f29d', '#dd48ab', '#E6FF80', '#1AFF33', '#32d9ce',
-		  '#ff80b0', '#CCCC00', '#66E64D', '#cacc4d', '#db45f5', 
-		  '#ffc4ef', '#2bff95', '#FF4D4D', '#99E6E6', '#6666FF'];
+
+// array of different colors
+let colorArray = ['#FF6633', '#ff8787', '#FF33FF', '#FFFF99', '#00B3E6','#f5c347', '#4dffea', '#ffc45e', '#99FF99', '#f26363', '#c3f24e', '#809900', '#61bf47', '#db8fac', '#b8e86f', '#5bfafc', '#CCFF1A', '#FF1A66', '#ffa599', '#33FFCC', '#66994D', '#f5c6b3', '#4D8000', '#B33300', '#CC80CC', '#eed4ff', '#991AFF', '#E666FF', '#4DB3FF', '#1AB399', '#E666B3', '#33991A', '#CC9999', '#B3B31A', '#00E680', '#f2f29d', '#dd48ab', '#E6FF80', '#1AFF33', '#32d9ce','#ff80b0', '#CCCC00', '#66E64D', '#cacc4d', '#db45f5', '#ffc4ef', '#2bff95', '#FF4D4D', '#99E6E6', '#6666FF'];
 
 
 let todoItemsContainer = document.getElementById("todoItemsContainer");
 let addTodoButton = document.getElementById("addTodoButton");
 let saveTodoButton = document.getElementById("saveTodoButton");
 
+// function to get the saved todoList from local storage
 function getTodoListFromLocalStorage() {
   let stringifiedTodoList = localStorage.getItem("todoList");
   let parsedTodoList = JSON.parse(stringifiedTodoList);
@@ -27,10 +21,12 @@ function getTodoListFromLocalStorage() {
 let todoList = getTodoListFromLocalStorage();
 let todosCount = todoList.length;
 
+// eventListener for SAVE button
 saveTodoButton.onclick = function() {
   localStorage.setItem("todoList", JSON.stringify(todoList));
 };
 
+//function to add a new todo task in the todo list
 function onAddTodo() {
   let userInputElement = document.getElementById("todoUserInput");
   let userInputValue = userInputElement.value;
@@ -52,15 +48,18 @@ function onAddTodo() {
   userInputElement.value = "";
 }
 
+// eventListener for ADD button
 addTodoButton.onclick = function() {
   onAddTodo();
 };
 
+//function to change the checked status of todo item when clicked on checkbox
 function onTodoStatusChange(checkboxId, labelId, todoId) {
   let checkboxElement = document.getElementById(checkboxId);
   let labelElement = document.getElementById(labelId);
   labelElement.classList.toggle("checked");
 
+  //getting index of todo item from todo list
   let todoObjectIndex = todoList.findIndex(function(eachTodo) {
     let eachTodoId = "todo" + eachTodo.uniqueNo;
 
@@ -71,6 +70,7 @@ function onTodoStatusChange(checkboxId, labelId, todoId) {
     }
   });
 
+  //change the status in todolist
   let todoObject = todoList[todoObjectIndex];
 
   if(todoObject.isChecked === true){
@@ -81,6 +81,7 @@ function onTodoStatusChange(checkboxId, labelId, todoId) {
 
 }
 
+//function to delete todo item when clicked on delete icon
 function onDeleteTodo(todoId) {
   let todoElement = document.getElementById(todoId);
   todoItemsContainer.removeChild(todoElement);
@@ -97,6 +98,7 @@ function onDeleteTodo(todoId) {
   todoList.splice(deleteElementIndex, 1);
 }
 
+//function to create and add todo item in HTML using DOM manipulation
 function createAndAppendTodo(todo) {
   let todoId = "todo" + todo.uniqueNo;
   let checkboxId = "checkbox" + todo.uniqueNo;
@@ -144,7 +146,10 @@ function createAndAppendTodo(todo) {
   deleteIcon.classList.add("far", "fa-trash-alt", "delete-icon");
 
   deleteIcon.onclick = function () {
-    onDeleteTodo(todoId);
+    let confirmed = confirm("Do You want to delete?");
+    if(confirmed === true){
+      onDeleteTodo(todoId);
+    }
   };
 
   deleteIconContainer.appendChild(deleteIcon);
